Feature: Add callbacks for filterfunctions to filter either a whole topmodule or specific submodules (or both)
The function `add_immediate_deps_to_modules` and `get_fq_immediate_deps` get two additional keyword arguments for
handing over callback functions that take module names and return a boolean value advertising if a module should be
processed (return true) or not (return false).

The lookup for these callback functions happens in a module called modfilter and the callback functions are named
`pkgfilterfunc` and `modfilterfunc` by default. If the module is not existent, standard behavior is assumed (no custom filtering).
If the module exists but neither `pkgfilterfunc` nor `modfilterfunc` are defined in the module, a notice is logged to stderr at the end.

For logging to stderr, a convenience function eprint is introduced.